마이페이지 테스트 코드 작성

+++ matgpt/src/main/java/com/ktc/matgpt/review/dto/ReviewRequest.java
@@ -14,12 +14,12 @@ public class ReviewRequest {
     @NoArgsConstructor
     public static class SimpleCreateDTO {
         private String content;
-        private double rating;
+        private int rating;
         private int peopleCount;
         private int totalPrice;
 
         @Builder
-        public SimpleCreateDTO(String content, double rating, int peopleCount, int totalPrice) {
+        public SimpleCreateDTO(String content, int rating, int peopleCount, int totalPrice) {
             this.content = content;
             this.rating = rating;
             this.peopleCount = peopleCount;
@@ -66,9 +66,9 @@ public class ReviewRequest {
                 private String name;
                 private double locationX;
                 private double locationY;
-                private double rating;
+                private int rating;
 
-                public TagDTO(String name, double locationX, double locationY, double rating) {
+                public TagDTO(String name, double locationX, double locationY, int rating) {
                     this.name = name;
                     this.locationX = locationX;
                     this.locationY = locationY;
@@ -112,7 +112,7 @@ public class ReviewRequest {
                 private String name;
                 private int locationX;
                 private int locationY;
-                private double rating;
+                private int rating;
             }
         }
     }
